# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Collaborators API",
        "version": "1.0"
    },
    "paths": {
        "/api/collaborators": {
            "get": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Display a listing of collaborators",
                "operationId": "3bad7de0da48224fdc7f177841872903",
                "responses": {
                    "200": {
                        "description": "List of all collaborators",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Collaborator"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Store a newly created collaborator",
                "operationId": "95299406c39c1533f68255ccedfe7d9b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the user",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Collaborator created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collaborator"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/collaborators/{id}": {
            "get": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Display the specified collaborator",
                "operationId": "b341e31bb3dfb9102075db15e330c455",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the collaborator",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Collaborator details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collaborator"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collaborator not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Update the specified collaborator",
                "operationId": "a640215bab0c29646ae2a95b3ebb2580",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the collaborator",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the user",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Collaborator updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collaborator"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Collaborator not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Collaborators"
                ],
                "summary": "Remove the specified collaborator",
                "operationId": "0c024bfecd29ff9123842fbb0c92de5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the collaborator",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Collaborator removed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Collaborator removed successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collaborator not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Collaborator": {
                "title": "Collaborator",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "post_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}